Move hardware specific code into module 'hal'
